Extracted mapping of client options to db options to a storage location
Cleaned up also dumbo imports
