+++ src/packages/pongo/src/e2e/compatibilityTest.e2e.spec.ts
@@ -1,4 +1,4 @@
-import { PostgreSQLConnectionString } from '@event-driven-io/dumbo/src/pg';
+import { PostgreSQLConnectionString } from '@event-driven-io/dumbo/pg';
 import {
   MongoDBContainer,
   type StartedMongoDBContainer,
